reorganizing test case files
