+++ cypress/e2e/WebInterface/Test Cases/QI-CORE Test Case/DeleteAllTestCases.cy.ts
