+++ cypress/e2e/WebInterface/Test Cases/QI-CORE Test Case/Execution/PassingCodeCoverageTestCaseListPage.cy.ts
@@ -1,13 +1,13 @@
-import { OktaLogin } from "../../../Shared/OktaLogin"
-import { CreateMeasurePage } from "../../../Shared/CreateMeasurePage"
-import { MeasuresPage } from "../../../Shared/MeasuresPage"
-import { EditMeasurePage } from "../../../Shared/EditMeasurePage"
-import { MeasureGroupPage } from "../../../Shared/MeasureGroupPage"
-import { Utilities } from "../../../Shared/Utilities"
-import { MeasureCQL } from "../../../Shared/MeasureCQL"
-import { TestCasesPage } from "../../../Shared/TestCasesPage"
-import { TestCaseJson } from "../../../Shared/TestCaseJson"
-import { CQLEditorPage } from "../../../Shared/CQLEditorPage"
+import { OktaLogin } from "../../../../../Shared/OktaLogin"
+import { CreateMeasurePage } from "../../../../../Shared/CreateMeasurePage"
+import { MeasuresPage } from "../../../../../Shared/MeasuresPage"
+import { EditMeasurePage } from "../../../../../Shared/EditMeasurePage"
+import { MeasureGroupPage } from "../../../../../Shared/MeasureGroupPage"
+import { Utilities } from "../../../../../Shared/Utilities"
+import { MeasureCQL } from "../../../../../Shared/MeasureCQL"
+import { TestCasesPage } from "../../../../../Shared/TestCasesPage"
+import { TestCaseJson } from "../../../../../Shared/TestCaseJson"
+import { CQLEditorPage } from "../../../../../Shared/CQLEditorPage"
 
 let measureName = 'TestMeasure' + Date.now()
 let CqlLibraryName = 'TestLibrary' + Date.now()
