+++ cypress/e2e/WebInterface/Test Cases/QI-CORE Test Case/Execution/RunAndExecuteTestCaseButtonValidations.cy.ts
@@ -1,13 +1,13 @@
-import { OktaLogin } from "../../../../Shared/OktaLogin"
-import { CreateMeasurePage } from "../../../../Shared/CreateMeasurePage"
-import { MeasuresPage } from "../../../../Shared/MeasuresPage"
-import { TestCasesPage } from "../../../../Shared/TestCasesPage"
-import { EditMeasurePage } from "../../../../Shared/EditMeasurePage"
-import { TestCaseJson } from "../../../../Shared/TestCaseJson"
-import { Utilities } from "../../../../Shared/Utilities"
-import { MeasureGroupPage } from "../../../../Shared/MeasureGroupPage"
-import { CQLEditorPage } from "../../../../Shared/CQLEditorPage"
-import { MeasureCQL } from "../../../../Shared/MeasureCQL"
+import { OktaLogin } from "../../../../../Shared/OktaLogin"
+import { CreateMeasurePage } from "../../../../../Shared/CreateMeasurePage"
+import { MeasuresPage } from "../../../../../Shared/MeasuresPage"
+import { TestCasesPage } from "../../../../../Shared/TestCasesPage"
+import { EditMeasurePage } from "../../../../../Shared/EditMeasurePage"
+import { TestCaseJson } from "../../../../../Shared/TestCaseJson"
+import { Utilities } from "../../../../../Shared/Utilities"
+import { MeasureGroupPage } from "../../../../../Shared/MeasureGroupPage"
+import { CQLEditorPage } from "../../../../../Shared/CQLEditorPage"
+import { MeasureCQL } from "../../../../../Shared/MeasureCQL"
 
 let measureName = 'TestMeasure' + Date.now()
 let CqlLibraryName = 'TestLibrary' + Date.now()
