+++ cypress/e2e/WebInterface/Test Cases/QI-CORE Test Case/TestCaseExportImport/TestCaseImportValidations.cy.ts
@@ -1,15 +1,15 @@
-import { OktaLogin } from "../../../../Shared/OktaLogin"
-import { CreateMeasurePage } from "../../../../Shared/CreateMeasurePage"
-import { MeasuresPage } from "../../../../Shared/MeasuresPage"
-import { TestCasesPage } from "../../../../Shared/TestCasesPage"
-import { EditMeasurePage } from "../../../../Shared/EditMeasurePage"
-import { TestCaseJson } from "../../../../Shared/TestCaseJson"
-import { Utilities } from "../../../../Shared/Utilities"
-import { MeasureGroupPage } from "../../../../Shared/MeasureGroupPage"
-import { CQLEditorPage } from "../../../../Shared/CQLEditorPage"
-import { MeasureCQL } from "../../../../Shared/MeasureCQL"
-import { Header } from "../../../../Shared/Header"
-import { Environment } from "../../../../Shared/Environment"
+import { OktaLogin } from "../../../../../Shared/OktaLogin"
+import { CreateMeasurePage } from "../../../../../Shared/CreateMeasurePage"
+import { MeasuresPage } from "../../../../../Shared/MeasuresPage"
+import { TestCasesPage } from "../../../../../Shared/TestCasesPage"
+import { EditMeasurePage } from "../../../../../Shared/EditMeasurePage"
+import { TestCaseJson } from "../../../../../Shared/TestCaseJson"
+import { Utilities } from "../../../../../Shared/Utilities"
+import { MeasureGroupPage } from "../../../../../Shared/MeasureGroupPage"
+import { CQLEditorPage } from "../../../../../Shared/CQLEditorPage"
+import { MeasureCQL } from "../../../../../Shared/MeasureCQL"
+import { Header } from "../../../../../Shared/Header"
+import { Environment } from "../../../../../Shared/Environment"
 
 let measureSharingAPIKey = Environment.credentials().measureSharing_API_Key
 let harpUserALT = Environment.credentials().harpUserALT
