+++ cypress/e2e/WebInterface/Test Cases/QI-CORE Test Case/TestCasePageCQL_PageObject.cy.ts
