+++ cypress/e2e/WebInterface/Test Cases/QI-CORE Test Case/Validations/TestCaseValidations.cy.ts
@@ -1,15 +1,15 @@
-import { OktaLogin } from "../../../../Shared/OktaLogin"
-import { CreateMeasurePage } from "../../../../Shared/CreateMeasurePage"
-import { MeasuresPage } from "../../../../Shared/MeasuresPage"
-import { EditMeasurePage } from "../../../../Shared/EditMeasurePage"
-import { MeasureGroupPage } from "../../../../Shared/MeasureGroupPage"
-import { Utilities } from "../../../../Shared/Utilities"
-import { MeasureCQL } from "../../../../Shared/MeasureCQL"
-import { TestCasesPage } from "../../../../Shared/TestCasesPage"
-import { TestCaseJson } from "../../../../Shared/TestCaseJson"
-import { Global } from "../../../../Shared/Global"
-import { CQLEditorPage } from "../../../../Shared/CQLEditorPage"
-import { Environment } from "../../../../Shared/Environment"
+import { OktaLogin } from "../../../../../Shared/OktaLogin"
+import { CreateMeasurePage } from "../../../../../Shared/CreateMeasurePage"
+import { MeasuresPage } from "../../../../../Shared/MeasuresPage"
+import { EditMeasurePage } from "../../../../../Shared/EditMeasurePage"
+import { MeasureGroupPage } from "../../../../../Shared/MeasureGroupPage"
+import { Utilities } from "../../../../../Shared/Utilities"
+import { MeasureCQL } from "../../../../../Shared/MeasureCQL"
+import { TestCasesPage } from "../../../../../Shared/TestCasesPage"
+import { TestCaseJson } from "../../../../../Shared/TestCaseJson"
+import { Global } from "../../../../../Shared/Global"
+import { CQLEditorPage } from "../../../../../Shared/CQLEditorPage"
+import { Environment } from "../../../../../Shared/Environment"
 
 let measureSharingAPIKey = Environment.credentials().measureSharing_API_Key
 let harpUserALT = Environment.credentials().harpUserALT
